+++ src/utils.ts
@@ -23,3 +23,26 @@ export const getDeferred = (): IDeferred => {
 
   return { promise, resolve, reject };
 }
+
+export const deepClone = function deepClone<T> (itemToBeCloned: T): T {
+  /* eslint-disable guard-for-in */
+  if (itemToBeCloned) {
+    if (Array.isArray(itemToBeCloned)) {
+      const clonedArray = [];
+      for (let i = 0; i < itemToBeCloned.length; i++) {
+        clonedArray[i] = deepClone(itemToBeCloned[i]);
+      }
+      return clonedArray as any as T;
+    }
+
+    if (typeof itemToBeCloned === 'object') {
+      const clonedObject = { ...itemToBeCloned };
+      for (const key in itemToBeCloned) {
+        clonedObject[key] = deepClone(itemToBeCloned[key]);
+      }
+      return clonedObject;
+    }
+  }
+  return itemToBeCloned;
+  /* eslint-enable guard-for-in */
+}